+++ ChangeLog.md
@@ -22,6 +22,8 @@ See docs/process.md for more on how version tagging works.
 -----
 - Updated SDL 2 port to upstream version 2.0.18 (from a patched 2.0.10). This
   includes all downstream patches and many upstream changes.
+- libc++, libc++-abi, and compiler-rt libraries have been updated to llvm-13.
+  (#15870)
 
 3.1.0 - 12/22/2021
 ------------------